Cleanup test utils for connecting to integrattions test DB
