+++ testutils/testutils.go
@@ -1,17 +1,9 @@
 package testutils
 
 import (
-	"context"
 	"fmt"
 	"os"
 	"reflect"
-	"testing"
-	"time"
-
-	"github.com/stretchr/testify/assert"
-	"go.mongodb.org/mongo-driver/mongo/options"
-
-	"go.mongodb.org/mongo-driver/mongo"
 )
 
 // SetEnvVars sets given environment variables and provides a callback function to restore the variables to their initial values
@@ -110,31 +102,4 @@ func (r RouterGroupMatcher) Matches(x interface{}) bool {
 
 func (r RouterGroupMatcher) String() string {
 	return fmt.Sprintf("router group's base path is %s", r.Path)
-}
-
-// ConnectToIntegrationTestDB waits for the integrations tests DB to become available
-// and returns a connection to the DB
-func ConnectToIntegrationTestDB(t *testing.T) *mongo.Database {
-	client, err := mongo.NewClient(options.Client().ApplyURI("mongodb://hs_auth:password123@localhost:8003/hs_auth"))
-	assert.NoError(t, err)
-
-	err = client.Connect(context.Background())
-	assert.NoError(t, err)
-
-	var db *mongo.Database
-	// Giving some time for the DB to boot up
-	for i := 0; i < 4; i++ {
-		db = client.Database("hs_auth")
-		err := client.Ping(context.Background(), nil)
-		if err == nil {
-			break
-		} else if i == 3 {
-			fmt.Println(err)
-			panic("could not connect to db")
-		}
-		fmt.Println("could not connect to database, will retry in a bit")
-		time.Sleep(5 * time.Second)
-	}
-
-	return db
 }
